SDK-1659: Refactor requests, move CreateShareUrl to YotiClient, merge shareurl package
Requests now made through internal request package, move CreateShareUrl to YotiClient, consolidate shareurl & dynamic_sharing_service packages
