+++ profile/activity_details_errors.go
@@ -1,4 +1,4 @@
-package yoti
+package profile
 
 import "errors"
